feat: Keep track of columns of issues, and all mispellings in file
While scanning the whole file looking for the error line now we also
find out which column the misspelling happened at. This has also the
side effect of finding other ocurrences of the misspelling in the file.

The Misspellings' offsets and Issues' columns are now being stored in
their respective ValueObjects.

This information can come in handy if we want to show where exactly the
misspelling happened (PR #28)[https://github.com/peckphp/peck/pull/28]
or if we decided to attempt to fix the file.